+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/message/CommonMediaGroupMessage.kt
@@ -1,6 +1,7 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.message
 
 import com.github.insanusmokrassar.TelegramBotAPI.types.*
+import com.github.insanusmokrassar.TelegramBotAPI.types.buttons.InlineKeyboardMarkup
 import com.github.insanusmokrassar.TelegramBotAPI.types.chat.abstracts.Chat
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.*
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.content.abstracts.MediaGroupContent
@@ -15,5 +16,6 @@ data class CommonMediaGroupMessage(
     override val content: MediaGroupContent,
     override val editDate: DateTime?,
     override val forwarded: ForwardedMessage?,
-    override val replyTo: Message?
+    override val replyTo: Message?,
+    override val replyMarkup: InlineKeyboardMarkup?
 ) : MediaGroupMessage, FromUserMessage