+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/message/abstracts/AbleToBeMarkedUp.kt
@@ -1,0 +1,7 @@
+package com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts
+
+import com.github.insanusmokrassar.TelegramBotAPI.types.buttons.InlineKeyboardMarkup
+
+interface AbleToBeMarkedUp {
+    val replyMarkup: InlineKeyboardMarkup?
+}