+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/buttons/InlineKeyboardButtons/SwitchInlineQueryCurrentChatInlineKeyboardButton.kt
@@ -1,11 +1,12 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.buttons.InlineKeyboardButtons
 
+import com.github.insanusmokrassar.TelegramBotAPI.types.switchInlineQueryCurrentChatField
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
 data class SwitchInlineQueryCurrentChatInlineKeyboardButton(
     override val text: String,
-    @SerialName("switch_inline_query_currentChat")
+    @SerialName(switchInlineQueryCurrentChatField)
     val switchInlineQueryCurrentChat: String
 ) : InlineKeyboardButton