+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/ChatIdentifier.kt
@@ -32,7 +32,7 @@ data class Username(
     }
 }
 
-fun String.asUsername(): Username = Username(this)
+fun String.toUsername(): Username = Username(this)
 
 @Serializer(ChatIdentifier::class)
 internal class ChatIdentifierSerializer: KSerializer<ChatIdentifier> {