+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/LoginURL.kt
@@ -1,0 +1,16 @@
+package com.github.insanusmokrassar.TelegramBotAPI.types
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class LoginURL(
+    @SerialName(urlField)
+    val url: String,
+    @SerialName(forwardTextField)
+    val forwardText: String? = null,
+    @SerialName(botUsernameField)
+    val botUsername: String? = null,
+    @SerialName(requestWriteAccessField)
+    val requestWriteAccess: Boolean? = null
+)