+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/buttons/InlineKeyboardButtons/URLInlineKeyboardButton.kt
@@ -1,9 +1,0 @@
-package com.github.insanusmokrassar.TelegramBotAPI.types.buttons.InlineKeyboardButtons
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class URLInlineKeyboardButton(
-    override val text: String,
-    val url: String
-) : InlineKeyboardButton