+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/message/CommonMessageImpl.kt
@@ -2,6 +2,7 @@ package com.github.insanusmokrassar.TelegramBotAPI.types.message
 
 import com.github.insanusmokrassar.TelegramBotAPI.types.MessageIdentifier
 import com.github.insanusmokrassar.TelegramBotAPI.types.User
+import com.github.insanusmokrassar.TelegramBotAPI.types.buttons.InlineKeyboardMarkup
 import com.github.insanusmokrassar.TelegramBotAPI.types.chat.abstracts.Chat
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.*
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.content.abstracts.MessageContent
@@ -17,5 +18,6 @@ data class CommonMessageImpl<T: MessageContent>(
     override val editDate: DateTime?,
     override val forwarded: ForwardedMessage?,
     override val replyTo: Message?,
+    override val replyMarkup: InlineKeyboardMarkup?,
     val paymentInfo: PaymentInfo?
 ) : Message, CommonMessage<T>, FromUserMessage