+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * `LoginURL` and `LoginURLInlineKeyboardButton` has been added
 * `replyMarkup` field was added to the `CommonMessage` objects via `AbleToBeMarkedUp` interface
+* `SwitchInlineQueryCurrentChatInlineKeyboardButton#switchInlineQueryCurrentChat` field fixed
 
 ## 0.15.0