+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * `replyMarkup` field was added to the `CommonMessage` objects via `AbleToBeMarkedUp` interface
 * `SwitchInlineQueryCurrentChatInlineKeyboardButton#switchInlineQueryCurrentChat` field fixed
 * `InlineKeyboardButton` now is sealed class and all its possible realisations are inside of its class file
+* `String#asUsername` method renamed to `String#toUsername`
 
 ## 0.15.0