+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * `SwitchInlineQueryCurrentChatInlineKeyboardButton#switchInlineQueryCurrentChat` field fixed
 * `InlineKeyboardButton` now is sealed class and all its possible realisations are inside of its class file
 * `String#asUsername` method renamed to `String#toUsername`
+* Several `toChatId` extensions added
 
 ## 0.15.0